Conflicts:
	src/site/root3287/sudo/screen/screens/OldTerrainScreen.java
	src/site/root3287/sudo/screen/screens/TerrainScreen.java
	src/site/root3287/sudo/terrain/perlin/PerlinTerrain.java
	src/site/root3287/sudo/terrain/perlin/PerlinWorld.java

+++ src/site/root3287/sudo/terrain/perlin/HeightGenerator.java
@@ -32,7 +32,7 @@ public class HeightGenerator {
     }
      
     //only works with POSITIVE gridX and gridZ values!
-    public HeightGenerator(int gridX, int gridZ, int vertexCount, int seed) {
+    public HeightGenerator(int gridX, int gridZ, int vertexCount, long seed) {
         this.seed = seed;
         xOffset = gridX * (vertexCount-1);
         zOffset = gridZ * (vertexCount-1);
